tools/design_tools.py
"""
Specialized Design Tools for DigiNativa AI Agents
==================================================

PURPOSE:
These tools give the Speldesigner agent cognitive abilities to validate
their own specifications against the project's core principles (DNA).
This ensures all design is high quality and aligned with project goals
before being handed over to development.

ADAPTATION GUIDE:
🔧 To adapt these tools:
1. Update the _run method in DesignPrinciplesValidatorTool to reference your own design principles
2. Adjust the prompt in AcceptanceCriteriaValidatorTool to match your quality requirements
"""
from typing import List, Type, Optional
import os
import json

# FIXED: Use correct import path for LangChain tools
from langchain_core.tools import BaseTool
from pydantic import BaseModel, Field
from langchain_anthropic import ChatAnthropic

# Import relevant constants from project settings
from config.settings import AGENT_CONFIG, SECRETS, DNA_DIR
from tools.file_tools import read_file

# --- Pydantic Input Models for Tools ---

class DesignReviewInput(BaseModel):
    """Input for DesignPrinciplesValidatorTool."""
    specification_text: str = Field(
        ..., 
        description="The complete text of the design specification to be reviewed."
    )

class AcceptanceCriteriaInput(BaseModel):
    """Input for AcceptanceCriteriaValidatorTool."""
    acceptance_criteria: List[str] = Field(
        ..., 
        description="A list of acceptance criteria to be validated.",
        alias="criteria_list"  # Allow both parameter names
    )

# --- Specialized Tools ---

class DesignPrinciplesValidatorTool(BaseTool):
    """
    A tool for reviewing a design specification against the project's
    five core design principles.
    """
    name: str = "Design Principles Reviewer"
    description: str = "Use this tool to validate a design specification against the five design principles. The tool provides scores and reasoning for each principle."
    args_schema: Type[BaseModel] = DesignReviewInput
    
    def __init__(self):
        super().__init__()
        # Initialize LLM for tool's internal logic
        self.claude_llm = self._create_claude_llm()

    def _create_claude_llm(self) -> Optional[ChatAnthropic]:
        """Create Claude LLM instance with error handling."""
        try:
            api_key = SECRETS.get("anthropic_api_key")
            if not api_key or api_key.startswith("[YOUR_"):
                print("⚠️  Warning: Anthropic API key not configured for design tools")
                return None
                
            return ChatAnthropic(
                model=AGENT_CONFIG["llm_model"],
                api_key=api_key,
                temperature=0.1,
                max_tokens_to_sample=2000
            )
        except Exception as e:
            print(f"⚠️  Could not initialize Claude LLM for design tools: {e}")
            return None

    def _run(self, specification_text: str) -> str:
        """Run the validation logic."""
        try:
            if not self.claude_llm:
                return json.dumps({
                    "error": "Claude LLM not available",
                    "fallback_review": "Manual review required - AI validation unavailable"
                })

            # Read current design principles directly from DNA document
            principles_path = DNA_DIR / "design_principles.md"
            
            if not principles_path.exists():
                return json.dumps({
                    "error": f"Design principles file not found: {principles_path}",
                    "suggestion": "Ensure design_principles.md exists in docs/dna/"
                })

            principles_content = read_file(str(principles_path), agent_name="design_validator")

            prompt = f"""
            You are a Senior UX Reviewer. Your task is to evaluate the following design specification
            against the project's 5 core principles.

            HERE ARE THE 5 DESIGN PRINCIPLES:
            ---
            {principles_content}
            ---

            HERE IS THE DESIGN SPECIFICATION TO REVIEW:
            ---
            {specification_text}
            ---

            Instructions:
            For each design principle, give a score from 1 (not followed at all) to 5 (perfect adherence).
            Provide a brief, concrete reasoning for your score.
            Respond ONLY with a JSON-formatted string following this schema:
            {{
                "principle_1_pedagogy": {{ "score": <int>, "reasoning": "<string>" }},
                "principle_2_policy_to_practice": {{ "score": <int>, "reasoning": "<string>" }},
                "principle_3_time_respect": {{ "score": <int>, "reasoning": "<string>" }},
                "principle_4_holistic_view": {{ "score": <int>, "reasoning": "<string>" }},
                "principle_5_intelligence_not_infantilization": {{ "score": <int>, "reasoning": "<string>" }}
            }}
            """

            response = self.claude_llm.invoke(prompt)
            return response.content
            
        except Exception as e:
            return json.dumps({
                "error": f"Design principles review failed: {str(e)}",
                "fallback_score": {
                    "principle_1_pedagogy": {"score": 3, "reasoning": "Manual review required"},
                    "principle_2_policy_to_practice": {"score": 3, "reasoning": "Manual review required"},
                    "principle_3_time_respect": {"score": 3, "reasoning": "Manual review required"},
                    "principle_4_holistic_view": {"score": 3, "reasoning": "Manual review required"},
                    "principle_5_intelligence_not_infantilization": {"score": 3, "reasoning": "Manual review required"}
                }
            })


class AcceptanceCriteriaValidatorTool(BaseTool):
    """
    A tool to ensure acceptance criteria are clear,
    testable and meet Definition of Done requirements.
    """
    name: str = "Acceptance Criteria Reviewer"
    description: str = "Use this tool to validate a list of acceptance criteria. The tool ensures each criterion is specific, measurable and testable."
    args_schema: Type[BaseModel] = AcceptanceCriteriaInput
    
    def __init__(self):
        super().__init__()
        self.claude_llm = self._create_claude_llm()

    def _create_claude_llm(self) -> Optional[ChatAnthropic]:
        """Create Claude LLM instance with error handling."""
        try:
            api_key = SECRETS.get("anthropic_api_key")
            if not api_key or api_key.startswith("[YOUR_"):
                print("⚠️  Warning: Anthropic API key not configured for acceptance criteria tools")
                return None
                
            return ChatAnthropic(
                model=AGENT_CONFIG["llm_model"],
                api_key=api_key,
                temperature=0.1,
                max_tokens_to_sample=2000
            )
        except Exception as e:
            print(f"⚠️  Could not initialize Claude LLM for acceptance criteria tools: {e}")
            return None

    def _run(self, acceptance_criteria: List[str]) -> str:
        """Run the validation logic."""
        try:
            if not self.claude_llm:
                # Fallback validation without AI
                fallback_results = []
                for criterion in acceptance_criteria:
                    fallback_results.append({
                        "criterion": criterion,
                        "is_testable": True,  # Assume testable for fallback
                        "suggestion_for_improvement": ""
                    })
                return json.dumps(fallback_results)

            prompt = f"""
            You are a QA Lead with expertise in writing testable requirements. Your task is to
            review the following list of acceptance criteria.

            LIST OF ACCEPTANCE CRITERIA:
            ---
            {json.dumps(acceptance_criteria, indent=2, ensure_ascii=False)}
            ---

            Instructions:
            For each criterion, evaluate if it is specific, measurable and testable.
            Respond ONLY with a JSON-formatted list of objects, where each object follows this schema:
            {{
                "criterion": "<the original criterion>",
                "is_testable": <boolean>,
                "suggestion_for_improvement": "<a concrete improvement if is_testable is false, otherwise empty string>"
            }}
            """
            
            response = self.claude_llm.invoke(prompt)
            return response.content
            
        except Exception as e:
            # Fallback validation
            fallback_results = []
            for criterion in acceptance_criteria:
                fallback_results.append({
                    "criterion": criterion,
                    "is_testable": True,
                    "suggestion_for_improvement": f"Manual review required due to error: {str(e)}"
                })
            return json.dumps(fallback_results)